jarvis/voice_jarvis/requirements.txt
requests
pygame
python-dotenv
